refatorando a main
